+++ guidata/tests/dataset/test_all_items.py
@@ -43,7 +43,10 @@ class Parameters(gds.DataSet):
     as well as special characters (α, β, γ, δ, ...)
     """
 
-    dir = gds.DirectoryItem("Directory", TEMPDIR)
+    dir = gds.DirectoryItem("Directory")
+    preview = gds.TextItem("File names preview")
+    option = gds.ChoiceItem("Option", (("1", "first choice"), ("2", "second choice")))
+
     fname = gds.FileOpenItem("Open file", ("csv", "eta"), FILE_CSV.name)
     fnames = gds.FilesOpenItem("Open files", "csv", FILE_CSV.name)
     fname_s = gds.FileSaveItem("Save file", "eta", FILE_ETA.name)